Validate single licence versions
We make the same kind of housekeeping changes we made to the licence validator. But like the fetch, we also split out validation of the purpose.

We also continue our approach to of working with single records rather than arrays of them. In this case validate just a single licence version.

The idea of validating that a licence has at least one licence version is a cracking idea. But with this new approach we know we'll have to build back the licence (attach purposes to licence versions, then the licence versions to the licence). We think with that done a distinct validator that then checks the structure of the whole licence will take the idea to the next level!
